+++ biosim/core/agent/Agent.java
@@ -1,5 +1,5 @@
 package biosim.core.agent;
 import biosim.core.body.Body;
 public interface Agent{
-	public double[] act(long step);
+	public double[] act(double step);
 }
